move sort-players file
